Add nonce check in plugin uninstall

+++ uninstall.php
@@ -2,16 +2,16 @@
 
 /**
  * Fired when the plugin is uninstalled.
+ *
  * @link       https://www.expresstechsoftwares.com/
  * @since      1.0.0
- *
  */
 
 // If uninstall not called from WordPress, then exit.
-if ( 
-		defined( 'WP_UNINSTALL_PLUGIN' ) 
+if ( defined( 'WP_UNINSTALL_PLUGIN' )
 		&& $_REQUEST['plugin'] == 'pmpro-discord/pmpro-discord.php'
 		&& $_REQUEST['slug'] == 'paid-memberships-pro-discord-add-on'
+	&& wp_verify_nonce( $_REQUEST['_ajax_nonce'], 'updates' )
   ) {
 			// Remove API credetials.
 			delete_option( 'ets_discord_client_id' );
@@ -36,14 +36,14 @@ if (
 
 			// Remove user meta
 			$ets_discord_user_ids = get_users( 'fields=ID' );
-			foreach ( $ets_discord_user_ids as $user_id ) {
-					delete_user_meta( $user_id, 'ets_discord_user_id' );
-					delete_user_meta( $user_id, 'ets_discord_access_token' );
-					delete_user_meta( $user_id, 'ets_discord_refresh_token' );
-					delete_user_meta( $user_id, 'ets_discord_role_id' );
-					delete_user_meta( $user_id, 'ets_discord_default_role_id' );
-					delete_user_meta( $user_id, 'ets_discord_username' );
-					delete_user_meta( $user_id, 'ets_discord_expires_in' );
-					delete_user_meta( $user_id, '_ets_pmpro_discord_join_date' );
-			}
+	foreach ( $ets_discord_user_ids as $user_id ) {
+			delete_user_meta( $user_id, 'ets_discord_user_id' );
+			delete_user_meta( $user_id, 'ets_discord_access_token' );
+			delete_user_meta( $user_id, 'ets_discord_refresh_token' );
+			delete_user_meta( $user_id, 'ets_discord_role_id' );
+			delete_user_meta( $user_id, 'ets_discord_default_role_id' );
+			delete_user_meta( $user_id, 'ets_discord_username' );
+			delete_user_meta( $user_id, 'ets_discord_expires_in' );
+			delete_user_meta( $user_id, '_ets_pmpro_discord_join_date' );
+	}
 }
